examples/document-detection/requirements.txt
ultralytics==8.* 
opencv-python 
numpy 
tqdm 
pyyaml
midv500
pyside6